feat: reorder files + make polymorphic iters
